pyproject.toml
[project]
name = "duode-tactical-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.11.7",
    "sqlmodel>=0.0.24",
    "pymysql>=1.1.1",
    "cryptography>=42.0.0",
]
